backend/multilabelling/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hyperframe==6.0.1
idna==3.6
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
minio==7.2.7
mkl==2021.4.0
modAL-python==0.4.2.1
mpmath==1.3.0
multidict==6.0.5
mutagen==1.47.0
mysql-connector-python==8.2.0
networkx==3.3
numpy==1.26.0
orjson==3.10.5
outcome==1.3.0.post0
pandas==2.2.2
pika==1.3.2
pillow==10.4.0
protobuf==4.21.12
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.19.0
sendgrid==6.11.0
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
skorch==0.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
starkbank-ecdsa==2.2.0
starlette==0.37.2
sympy==1.12.1
synchronicity==0.6.7
tabulate==0.9.0
tbb==2021.13.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.1.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
yt-dlp==2024.5.26
